requirements.txt
PyYAML~=5.3.1
emoji~=0.6.0
numpy~=1.19.0
discord.py~=1.6.0
typeguard~=2.9.1
loguru~=0.5.2
sqlitedict~=1.7.0
pytz~=2020.5